+++ example_test.go
@@ -60,7 +60,7 @@ func Example() {
 
 	m := runnable.NewManager()
 	m.RegisterService(jobs)
-	m.RegisterProcess(runnable.Schedule(cleanup, runnable.Every(500*time.Millisecond)))
+	m.RegisterProcess(runnable.NewSchedule(cleanup, runnable.Every(500*time.Millisecond)))
 	m.RegisterProcess(runnable.Named(runnable.Func(func(_ context.Context) error {
 		<-done
 		return nil